+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Compatibility with Zygor Guides Viewer for the Blizzard Achievements tab
 - Focused, Current Zone, Selected Zone and Excluded category to every tab (option to hide them in unwanted tabs to be added)
+- Refresh to World and Widget Event Reminder options
 
 ### Fixed
 - When an achievement is selected and a criteria is completed, the criteria should show as completed without having to toggle the achievement